FirmwareUpdater V1.3: using pin PA4 for SD SPI CS. added missing __FIRMWARE to `st_clust()`. maybe we need an automated static analysis to verify all functions and global/static variables conform in the required files? using PA1 for random values instead of PA4 (as PA1 is connected to the DETECT pin but can sill be used as it has sufficient noise). changed all doubles to floats for better memory management (floats are good enough for the precisions we need). added support for the BNO085. also fixed bug for the wire library so that if you fill it up completely it doesn't think it is empty

+++ examples/si1145_test.cpp
@@ -18,9 +18,9 @@ void setup() {
 }
 
 void loop() {
-    const double uv_index = uv_sensor.read_uv_index();
-    const double visible_lux = visible_lux_sensor.read_visible_lux();
-    const double ir_lux = ir_lux_sensor.read_ir_lux();
+    const float uv_index = uv_sensor.read_uv_index();
+    const float visible_lux = visible_lux_sensor.read_visible_lux();
+    const float ir_lux = ir_lux_sensor.read_ir_lux();
 
     PAL_SERIAL.print("UV Index: ");
     PAL_SERIAL.print(uv_index, 2); // 2 d.p.
