+++ lib/interfaces/HumiditySensor.h
@@ -9,8 +9,8 @@ class HumiditySensor {
         /**
          * @brief Read the humidity.
          * 
-         * @returns A double of the value of the humidity in %RH.
+         * @returns A float of the value of the humidity in %RH.
          */
-        virtual double read_humidity() = 0;
+        virtual float read_humidity() = 0;
         virtual ~HumiditySensor() = default;
 };
